refactor(): Move App.vue to components/ dir
