+++ resources/js/components/App.vue
